The beginning of a notifier system & DAP interface
The scaffolding for the DAP user interface is now in place. This is
getting pulled in, because I think it's safe to assume that iterating on
this will involve filling out the remaining functionality without making
too many major changes to the scaffolding (the UICommand/UIResult event
queue model).

Three threads in effect;
Tracer thread - ultimately executes all user commands, performs debug info symbolication and all the things related to a debugging experience.

The Awaiter thread - a task with a sole purpose; do a blocking waitid (VERY IMPORTANT; not waitPID but waitID, waitpid doesn't support WNOWAIT! Which fucking SUCKS!). When it sees that a child has changed it's state it notifies the Tracer thread that it can do a waitPID.

The IO Thread - This thread reads input/output from the tracee as well as the request stream that comes from the client (via stdin).

Tracer thread communicates with all threads, whereas the other threads only communicates with Tracer.