StackTrace Request
Added the reading in of all dies in CU.
Record all subprogram dies

This was a first-hack approach in order to get the debug info data that I need. Going forward,
one would have to think hard to not have to do repetitive iterations over the debug symbols. But that's a problem for another day.