Get register value using dwarf reg num
Computes the offset for a given register in user_regs_struct
and retrives the value that way, instead of having to keep a std::map
of number -> reg.<register_field> or some switch/if-statement.

Sure, this can be fragile, but it's a lot faster and a lot easier.
It's also very easily unit tested (see future parse_cf gtest/unit test)