Unwinder needs ObjFile*
It requires this to (at a later point, haven't implemented it yet)
resolve a possibly relocated address of the .text section in that
object file.

For instance if a shared library is loaded, the .eh_frame might point to
some weirdo address - when in fact this is used to create a new address
(relocation). We need to understand this and so does Unwinder otherwise
it will be reporting the address that `readelf -S foo.so` reports for
the text section, i.e. prior to relocation.