Re-factor Task's bpstat
It's now been moved into TaskInfo to be managed explicitly by the task info object itself.

There was no reason to coalesce all these into a vector in breakpoint map, as it just becomes
doing a bunch of lookups all the time; now it is accessible directly within the TaskInfo object. Sure, the TaskInfo object now grows with 16 bytes in size, but seeing as how this operation is done on *every* stop, I think it's worth it.