Waitstatus processing re-factor A major re-factor of processing of wait status in order to simplify future work, where for instance the awaiter thread continuously waits and pushes it's results to a global main event queue.
Fixed a "logical bug" in the sense that user interface notification happens too early (before MDB has processed all the state it needs).

For instance; if a thread triggers a breakpoint, and all other threads are to be stopped, the DAP "stopped event" (in this case, but in the future, when possibly new UI:s are added, this will also be true) needs to be sent after _all_ threads have been stopped and their wait statuses processed, not immediately when the breakpoint event happens. Instead, sending a DAP "stopped event" should only immediately happen if _only_ that thread is to be stopped.  TraceeController now has a flag `waiting_for_all_stopped` which is checked in `StopHandler::handle_wait_event`; when all are stopped, it empties the event buffer, stored in the `StopObserver` a new type introduced in events/event.h

Instead of resuming by using TaskInfo::ptrace_resume  directly, all resuming should go through via TraceeController::resume_task(TaskInfo*, RunType), because this makes sure that all state, with respect to TaskInfo is propely set/configured.