First enumerate view implementation
Adds an index to a range based for loop, so we don't have to manually track it (and potentially screw it up).

When analyzed in godbolt, there's a ~7 instruction in total overhead, possibly involving the destruction of the temporary
EnumerateView object - though I'm not sure. The actual iteration instructions are almost 1-to-1 (2 instructions that differs)