First stage commit of new DWARF interpretation
- Added the ability to parse compilation unit & die meta data in parallell, using `UnitDataTask` & `IndexingTask`,
where `IndexingTask` maps names to a die index. Due to the design of how MDB maps out dies, this is a direct index, that is stable, no matter
if the DIE structure is loaded or not (because it will be the same, every time it's loaded). This also gives us the possibility to load dies of a CU, and then unload the layout metadata, if we want to save space. This is a future problem to deal with though.

- Added `calculate_job_sizes` as a static member function on `ThreadPool`; it takes a span of T, and calculates how it can divide those items up on the total amount of threads in the thread pool

- Added UnitReader that models the old one almost 1-to-1 with additional features. Not fully complete, needs to be able to do the resolving inter-die-reference stuff. (ugh).