Parsing of DWARF using new system on Shared Objects
Parsing of DWARF using the new system is moving along nicely.

Currently, shared library objects are parsed *only* using this feature.
But as function symbols aren't yet generated using this meta data, on test fails,
the one that unwinds the stack from a shared library (libmetricsconv.so).

That fix will come in the next commit. After that, we should be able to
remove _all_ the old code involving DWARF DIE-data.