Fixed bug that only showed up during gcc-built targets
The bug is still a bug using clang though, it just hadn't shown up yet.
We need to make sure that when searching for an address and line table entry, that the "found" lte
actually belongs to the frame we are interested in.

Same goes for get_fn_by_pc; it should search end exclusive, as one frame might have an end pc == start pc of another frame.