Big Re-factor & Type Parsing
Removed excessive pointer usage: use references instead.
New types have been introduced where it doesn't make sense to use references
in these cases we use `NonNullPtr` instead. This can be, because maybe we want an "identity"
function for an actual object. This doesn't work with references.

Added the types `NonNullPtr<T>` and `Immutable<T>` - these are more or less self explanatory.
It makes it easier to reason about types and how they're used. It also makes it possible
to more expressively use `struct Foo { ... }` since we are no longer concerned about stupid getters
or setters, or a factorial explosion of constructors etc. we get value semantics for free! std::move
works automatically (most often, that is)!

Increased parse buffer size to 16 pages (4096 * 16), mimicking linux's default pipe size.

## Type Parsing
This is the initial work of parsing & resolving types from the DWARF debug symbol data.
I'm actually quite ok with the general idea so far. In the future we might think about adding
a parallellization of the process, so that when we parse symbol information for frame at level N
we also parse it in parallell on levels N+1 .. N+M; alternatively, we parse a deeper depth
of the types in the current scope, in parallell (ideas are not mututally exclusive).