Tests, Variable Visualization, Args validation
Visualization & Value Resolving

Added first steps of infrastructure for data visualization. Currently
the idea has landed in having two distinct concepts working together
towards displaying values: Value Resolver and Value Visualizer. The
value resolver types can take a `Value` and produce new (possibly)
values or sub-values from that one. For instance, in the case of a value
that is a `const char*` there is the built in CStringResolver, which
reads the pointer value, and "resolves" the data it points to into a new
Value. This new value has a CStringVisualizer registered with it, which
handles the logic of how to format this resolved value into the value
field of a DAP variablesRequest.

Thus, we have a "hierarchy" here, in how Midas Debugger thinks of
values:

MemoryContentsObject: The actual bytes contained at [address .. +length]

Value: The intepretation of [n:m] of MemoryContentsObject

ValueResolver: A value resolver that from a `Value` produces new
`Value`s. Those values can be sub objects (like for instance a member
variable of a struct), in that case, they (re)use the same
MemoryContentsObject as the parent value, but with the appropriate
range. But it may also create pointee values, which means we must create
a new MemoryContentsObject for where the pointer pointed to.

ValueVisualizer: The custom visualizer for data that handles the logic
of how to write specific `Value` to a string (the value field in DAP)

Tests
Added additional tests and cleaned up considerably.

Command arg validation
Simple args validation has been added so that the DAP command request
parser doesn't assert on every badly formatted input. Has only been
tested very little (only a few tests in the test dir test faulty args.
the readMemoryRequest suite has a test for faulty args. *Every* test
suite should have a test for faulty args for it's corresponding request
it's testing)

	#M