Value Object Cache Refactor
Instead of tying individual values and value objects to a `SymboFile` (which in turn, is unique for a Supervisor/TraceeController), the value cache is now instead inside the `TaskInfo` class.

Why? Because that way, we can more smoothly support single thread / all-thread control. We don't want to invalidate thread state for a thread that is not currently executing. That would be wasteful.

Bug fixes:
- Remove of breakpoint location user segfaulted for UserBreakpoint with no bploc
- `ThreadPool` (the global one) now has shut down functionality added so that we can gravefully quit.