Bug fix & feature add
Fixed bug when type is a typedef, MDB did not treat it as the resolved type during Variables requests, thus failing to report, e.g. an array of N elements ( _M_elements in std::array is a typedef of a raw C-array)

Added the feature of resolving values behind a pointer ("chasing a pointer"). So now VSCode can display the actual pointee instead of just the pointer value.