Feature fix
Resolve references to the "next" type, instead of the fundamental layout type.

Why? Because if we have a pointer to a pointer, we don't want the ultimate pointed-to-type (what we here at MDB call the fundamental layout or fundamental type), we want the "next" type.

so int**, we want int*. And resolving int* will give int.

But both int** and int* share `int` as it's fundamental type. Because pointers are just 8 byte values, really. They're _not_ "real" types.