Initial work to get `fork` support done
The following patches will involve refactoring out the DAP class, to instead of being a global once-per-Tracer instance (which is  a global singleton) we will now spawn 1 DAP per supervisor instead, because the DAP, in it's infinite (infintessimal, more like it) wisdom, decided to not just send a `pid` id with the requests. And the usual suspects are against everything that would rock in the protocol. But oh well.