Fix support for plt-entries | relocatable code
Tries sniffing out a frame that's just a PLT entry by reading what's in the address [$rsp] (i.e. rsp holding a pointer value to an address in memory) because the `call` instructions always pushes the return address to $rsp (and plt entries are called via `call` at least on linux... so that'll be fine).

- Reset of call stack change
- clang-format