Bug fix - not handling bad remote reads
We didn't handle the case where a dereference of a pointer in tracee address space, that points to garbage memory, results in no value.

We should probably at some point introduce `InvalidValue` as a composed type containin a Value or a sub type of value - either or, i haven't decided yet.

For now, it's just a `Value` that uses it's intended `InvalidValueVisualizer` to "display" (it's really just serialization for the debug adapter protocol format).

Re-factored common.h so that it no longer contains TraceePointer<T>; it has instead been moved to it's own file tracee_pointer.h

N.B: Additional clean ups should follow from this, so that compile times is as optimized as possible, i.e. remove any dead includes. Since I'm in the process of cleaning up, it should be in one of the following commits, if no new bugs are introduced while use-testing.