Initial support: allocators in command subsystem
Added std::pmr::memory_resource* function arguments to functions and command objects
making it possible to allocate from a temporary arena, that can be reset
once the command & it's result has been output.

Future work will involve adding a similar system to the
Variables command, where each "stop" will have a sort of
"Frame" allocator, sort of like a game engine.

This frame allocator, will hold *everything* that is allocated
between a stop (e.g. a StoppedEvent being sent to the debug adapter)
and a Continue request; because between these two points in
debugger-time, the tracee will not have changed - and even if it did
it would be trivial, I believe, to either copy data around
or mutate it, so long as it belongs to the current "frame".