Bug fix in TypeStorage
Type storage in ObjectFile was not created with TypeStorage::Create, so the void type was not installed by default (thus storage[0] pointed to a nullptr instead pf a type).

SymbolFile::GetVariables now handles properly frames that does not have full debug symbol infor (and just return an empty list, instead of asserting).