Some "data types" will actually have a size of 0.
What does this mean? Well a pointer to a subroutine type, (i.e. a function pointer) will point to a T, which in C/C++ is a function, and functions do not have type size. The pointer itself is obviously easy: it's 4 or 8 bytes depending on if you're running 32 bits or 64 bit address space. So it will always be 8 bytes, but trying to "derefernce" a pointer to a type of size 0, like for instance subroutine, but also a pointer to void, will result in not an error but a fail-to-read, which is perfectly fine.

It will be trivial in the future, to add operations that re-interpret void* to, say buf: [u8; 1024], if we "get to know" that a void pointer actually points to a buffer of 1024 bytes. Meaning, in the future it should be possible to write pretty printers that re-intepret a void* in some type T to a pointer to a buffer (if that's what the type does). Hard to explain and probably harder to understand without context.