Pre-refactors
- Added logging to various functions
- Removed process_group field on the Event for the main debugger event loop type, there's no point for it.
- Added VFork support
- Clone child bugs addressed (thought not 100% sure it's fixed)
Some children that when cloned from their parent have their wait status reported to us before we've been informed by the cloner that it actually cloned itself. This introduces logic & timing issues. To address this, I've added a "UnInitialized" kind task, that once the clone event from the cloner *actually* comes through, which it always should, we push a core event "ThreadCreated" to the core debugger loop, which will come after dealing with the cloner-stop, and resume the newly minted clonee.