Added test for new source code file refactor
When compiling tests, we have person.h as a `SourceCodeFile` that *should* be created once for each compilation unit, a.cpp and b.cpp, whereas before I tried to merge these. But I quickly found that there was absolutely no benefits with that approach except in the *extremely* narrow use case where one want to set breakpoints on `filepath.h:1234`. But this use case happens once every 15 years, for what it's worth for some files, while looking up a line table entry in a compilation unit and quickly mapping it to source code file happens literally *every single stop*, so, hundreds of thousands of times.