Source Code File refactor
Now every compilation unit creates a `SourceCodeFile` for each file in it's LNP header.  Prior to these commits
we would merge identical paths in different CU's. This just caused issues, since it made binary searching large line tables harder.

Instead the interface is something like this;
You have a `SourceCodeFile` object, that maps to some file foo.h or foo.cpp in a compilation unit; in order to retrieve it's *exclusve* line table entries from that CU, one can use `SourceCodeFile::ReadInSourceCodeLineTable`  into a vector of line table entries and then search that, if one want to do an operation that involves a source code file and some source code coordinate, e.g. setting a breakpoint att foo.h:123

These changes come with substantial changes throughout the system as this is one of the core design items (the SourceCodeFile, that is). It affects how breakpoints get their souce code location info, like what file and line they may be set on, what file a type is declared in and so on.

Since we've merged the SourceCodeFile, it also no longer means that if a type is defiend in Foo.h; we don't have 1 source of truth, but possibly multiple; however, as far as the type interpretation system goes, we link the different die's to the same `sym::Type` pointer, to avoid having to parse types many times, this works as first-come-first-served, meaning, if a DIE in a compilation unit somewhere has a type and it gets parsed and evaluated fully, if anoter compilation unit wants to parse that same type, it will add it's DIE offset to the type, and check if that type exists already. That way, multiple DIE's can refer to the same MDB-type.