Refactor of AddrRange -> CompUnit mapping
Now we map an address range to sym::CompilationUnit* instead of UnitData*, as the UnitData* really is just a handle to the actual bytes with some metadata, it's the sym::CompilationUnit* that's the interesting parts.

It also speeds up, substantially, where we want to look up a CompilationUnit* by address key.

Also handles buggy dwarf AddressRange data in .debug_aranges/.debug_ranges where ranges can be 0x1 -> 0x1, which causes some compilation units to have a much larger address space, which resulted in 1500 matches for a stack trace in firefox. With the fix, it's down to 150 ish, but there's additional room for optimization here, I think.