Refactored event queue
This now is signal-safe, since we write the waitpid results over the pipe and serializing the wait events over a file descriptor is safer than attempting to take a lock, which *can* be interrupted by a signal, and attempt to re-take a lock for a new event, which dead locks.

Also added the UnInitialized tasks and some minor other bug fixes.