Added /* LICENSE TEMPLATE */
Sooner or later (sooner rather than later) a license will be determined
for this debugger. There's not a zero-chance that the license will
not be entirely free.

When license is picked, it will be fairly trivial to generate it
in-place in the source files that have this comment at the top.