Added DebuggerThread
DebuggerThread is a light wrapper around a std::thread
that also makes sure that SIGCHLD as a signal is blocked before
spawning a thread as well as after it having started.

The idea is that we are going to move away from a signal-waitpid-routine
to a epoll-signalfd-waitpid routine, thus one that does not use a signal
handler for SIGCHLD, to increase performance (drastically) for
application that creates many child tasks.