Unfortunate commit
This had to be done right now. This is catastrophic re-factor, but it just needed to be done now so that one can have a glimmer of hope of not trashing the version history, going forward.

It has some stuff that will be removed in the next few patches, that is just here for testing