Refactored `Value` type
The previous system was too convoluted. Moved out visualizers and resolvers and simplified the code substantially. We don't try to hold intra-value references and instead create copies, because there are so many problems and hurdles to tackle to make that work, that it's not worth focusing on right now, what's more important is moving on to the scripting and get that ball rolling and eventually do profiling if we hit some performance bump (because I don't foresee it being a bottle neck, but we do want to eek out all the performance if we can).

Added Weak reference ref counted types, using a control block as well. `sym::Value` is of thise ref counted type, because I think there is _some_ intra-value reference holding that might be worth it, and for these scenarios weak reference sharing is the only possible way, since Javascript may hold a reference to a sub object of a value.