Added many new JS API functions
- Create framejs, variablejs
- Updated taskinfojs, util, bpjs

Refactored variable references to better support the new JS API.

Because of this, any previous caching has been removed. Caching will have to consider both debugger-core as well as debugger-js domains. Therefore sym::Value has been made Ref counted-with-weak-ref support.

The idea I'm thinking about is having a "variables reference boundary/stop". Then we can compare any value, on the javascript side and check "if v.id() > mdb.varRefBoundary()" then it's still valid (because at a new stop a new boundary would be set).

This way, we can keep values alive for as long as we want to, and we will have a fairly easy way to know, which one "can't be trusted" so to speak (and thus have to be refetched or "refreshed" - another idea I think about implementing).