Removed eval/folder & Begin JS Repl
This is because going forward, Javascript / Spidermonkey will be the evaluator, so no point in creating an ad hoc scripting language.

Implements a couple of additional global functions, that the user
can use to query for state in the debug session. The repl will work sort
of like a normal js repl. So to get a task we'll do: getTask(12345 /*
pid */), but the user must also know what tasks there are so functions
like that needs creating too.