Added sessionId's to supervisor and threads
These session id's are debugger id's which is debugger local. The id's are monotonically increasing and always a unique. there's one id for the supervisor and one for the tasks, so supervisors and tasks both start at id 1. This is a convenience id, for uses that want to access a supervisor or task using it's (often shorter) session id. The debugger core itself though, always uses the pid_t from the kernel to identify a process or thread.