Add AllStopSession flag & initial support for event time
Event time reflects a monotonic counter. When debugging a replay session, RR will notify us of this value during reverse execution or forward execution. That way our core debugger loop doesn't need to change all that much - we will know what trace time it is and handle it accordingly, if we've reversed across the birth of a process or thread.

This is just scaffolding work in this commit.