Extend new build system
Added new commands:
- build
- clean
- dev-setup
- configure-buildroot
- configure
- list-presets
- select
- autocomplete

The first command that always must be run is `dev-setup` as this will take care of the dependencies going forward, instead of the bash script. It has the added benefit of downloading the Spidermonkey source code. It's huge. It builds both a debug and a release version and puts them in the dependencies folder. The huge size probably will make me not want to use it in the future. But we'll see.

build can build one of the configured presets. The presets have to be configured using the `configure` command. mdbuild list-presets lists the cmake presets that can be built.

`configure-buildroot` will configure the root folder where the build directories for the presets can be set in.

`autocomplete` creates a bash script that can be sourced so that the tool can be auto completed from the command line. Add the source folder path to $PATH and mdbuild can be used without the ./ before.