Style guide fixups & Restructuring
The (pretty huge) cleanup that needs to happen as far as structuring of
the code, and being consistent with styles will have to happen in
stages, every now and then.

Moved implmenters of UICommand and UIResult to commands.cpp as these are
not needed to be exposed via a header/interface. The interface *is* the
UICommand/UIResult.