Add command line parser
The command line parser takes references to the values where it's supposed to store the parsed data.

All values need a default value set, so that for arguments that are not passed in, they have a fallback. In the future, we can add required arguments.

Following arguments added in this PR:
-h, --help

-i, --interface: takes either "stdio" as input, or "/tmp/some-unix-socket-path" - note that mdb will only ever accept files that live under /tmp. The prefix unix:/tmp/foo also works; this is future proofing, when we may want tcp as well. It's over this file or socket, that we communicate using the debug adapter protocol. Thus, a client can pass in a socket here, that mdb will connect to and communicate it's DAP over. DebugAdapterClient::CreateSocketConnection is where most the logic for setting this up is. This is not fully tested yet and has not been pulled into the test suite.

-w, --timeout: the amount of time mdb will wait until it's timed out when attempting to connect to the socket provided by the user in -i/--interface

-l, --log: the directory where the log files should be stored, defaults to current working directory

-t, --threads: configure the thread pool size/the amount of worker threads to be used.

Environment variables that are parsed using the same system:

LOG=<channelA>,<channelB>, ... see the help message (mdb -h) for more information.