Add CreateRequest for Attach and Launch DAP requests
Moved the logic for creating a LaunchRequest into its CreateRequest function. This is because going forward, we'll add arguments, so that the user can specify that it's launching a trace, and as such needs extra args.

All requests will eventually have their own `CreateRequest` static function that parses the incoming request parameters, so that error handling etc, isn't all handled in `ParseDebugAdapterCommand` but in the respective create request function.