Cleaned up responsibilities for `Tracer`
The tracer singleton is just a "global object", containing handy pointers to data of varioius kinds. It's supposed to be that, too.

Therefore removed `ForkExec` and friends from this interface, as it belongs to `PtraceCommander` instead, since it's the only tracee interface that actually spawns new processes (from the debugger's perspective). Processes spawned by RR in the future, is of no concern to mdb.

There are a few things like this that needs refactors. Particularly, I believe that all implementations of TraceeInterface actually needs their own event loop, not this loosely glued/held together thing that currently exists in `TraceeController::DefaultHandler` which *very often* run into edge cases that depends solely on what interface it is using. This is a big clue that a redesign there needs to happen.