Resume & Scheduling refactor
This commit refactors how scheduling of tasks work and who is responsible for who. Not happy with the final result and a re-write (meaning, not an iterative add-on) will probably still be in the future. But this commit is less error prone and has been tested to work against remote targets, ptrace as well as the soon-to-be-introduced rr-target which is hosted natively within the debugger.