Added `control` as a log channel & some logging statements
Renamed mInitialized to mHasStarted to reflect that newly created threads has been resumed at least once. This is to deal with when sending signals to the thread group; targeting un-started threads is pointless. better just leave them as is.

This commit also sneaks in DBGBUFLOG, which formats to a stack-local buffer of 1024 bytes instead of allocating a string.